Break Comm232 into its own file
